Add lots of changes (See Description)
Improved the tab list with abbreviated rank prefixes and coloured names. Also added a header with the name and a footer showing the number of online players.
Added a rudimentary town command to set up/leave a town. Also added a rudimentary money system.

Changed role colours to look nicer.